src/main/webapp/app/entities/despliegue-infraestructura-troncal-distribucion/delete/despliegue-infraestructura-troncal-distribucion-delete-dialog.component.ts
import { Component } from '@angular/core';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { IDespliegueInfraestructuraTroncalDistribucion } from '../despliegue-infraestructura-troncal-distribucion.model';
import { DespliegueInfraestructuraTroncalDistribucionService } from '../service/despliegue-infraestructura-troncal-distribucion.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  templateUrl: './despliegue-infraestructura-troncal-distribucion-delete-dialog.component.html',
})
export class DespliegueInfraestructuraTroncalDistribucionDeleteDialogComponent {
  despliegueInfraestructuraTroncalDistribucion?: IDespliegueInfraestructuraTroncalDistribucion;

  constructor(
    protected despliegueInfraestructuraTroncalDistribucionService: DespliegueInfraestructuraTroncalDistribucionService,
    protected activeModal: NgbActiveModal
  ) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.despliegueInfraestructuraTroncalDistribucionService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
